requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
crispy-bootstrap5==2024.2
Django==4.2.14
django-bootstrap-v5==1.0.11
django-crispy-forms==2.2
django-phonenumber-field==8.0.0
docopt==0.6.2
et-xmlfile==1.1.0
num2words==0.5.13
numpy==2.0.0
openpyxl==3.1.5
pandas==2.2.2
phonenumbers==8.13.40
phonenumberslite==8.13.40
pillow==10.4.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
tzdata==2024.1
